requirements.txt
flask
numpy
pillow
tflite-runtime==2.13.0
gunicorn